Fixed bug in ColorSeries class and in IconLegend that were leading to segmentation fault.
A new bug was manifesting in Madeline when an AFFECTED column populated solely by missing
values was included in a group of AFFECTED columns treated as the ICON columns on a pedigree.

This bug did not manifest if the same AFFECTED column included at least one non-missing value.

The bug ultimately was due to the fact that the ColorSeries had originally been written with
scant support for handling missing and out-of-range values. I have now rewritten much of the
ColorSeries class so that it now includes a static private DrawingColor _missingColor which can
be returned as a sentry (or whose methods can be called as needed). Also the private but
redundant _levels variable has been removed, since the private std::vector<_colorSeries> already
provides size().