Push custom colors in reverse order
In order to get the results that the user expects,
custom colors must at some point be ordered in reverse.

In this commit, during command-line processing in main.cpp,
we add a temporary std::vector<std::string> and then use a
reverse iterator to push the colors onto the ColorSeries object.

Although this 'reversing' could be done elsewhere in the code,
it is easy to do it here. Note that this might change if I decide to
overhaul the color vs. black-and-white decision-making machinery
in Madeline at some future point in time.