Added customColorMap to DrawingMetrics
Recently a customColorSeries was added to DrawingMetrics to permit
the specification of custom colors.  The code as written implicitly
assumes that for a given pedigree presented as a data set, all levels
of a displayed categorical variable will be present. However this can
lead to faulty output if one or more levels of a categorical variable
happen to be missing from the data set because the vector implementation
just maps the ordinals rather than the actual labels.

Using a map between specific labels and colors of a categorical variable
would eliminate this problem. This commit therefore takes the first steps
toward implementing a custom color map. Note that since there can be
multiple categorical variables displayed on pedigree drawing, the actual
implementation is a std::vector of std::map< std::string,std::string>  .

The implementation is not yet complete: however, there should be no
regressions of the current functionality at this point in time.