Add recognition of set of 140 named colors
The DrawingColor class now recognizes the set of
140 English color names that are, for better or worse,
"standard" in all modern browsers (inter alia?). These named colors
originated with the X Windowing system although a
few names differ in the W3C standard. I am *_not_* a particular
fan of this set as I feel that there are a fair number of names among the
140 that will be completely unknown to almost everyone, even
among native speakers of English. Nevertheless, the set includes
widely-recognized basic color names such as "red", "black", "magenta",
etc. as well as commonly recognized color names such as "salmon"
and "lime". I may in fact cut out some of the insanely meaningless
names such as "NajahoWhite" which is more of some sort of adobe color;
or PapayaWhip which is ... I seriously have no idea what color that is!!!

In any case, additional work is required: this is a work-in-progress.