IconLegend now prints level colors from color map
The icon legend now correctly prints the level colors
when a custom color map is used (i.e., when colors are
specified using syntax such as this:

  -C "NORMAL=#4f4,OBESE=#f44,OVERWT=#ffc800,UNDERWT=#98d6ff"

Please note however that some additional refactoring is
still needed. In particular, the code to determine the ink color
to use is not correct at this point (Because of other commitments
today, I'll fix that tomorrow).