Fixed segfault on bad original founder data
The program was segfaulting on a bad data file where a pair of
original founders were both missing for the GENDER attribute.

Although we could probably code Madeline to auto-correct this,
based on knowledge of which one of the pair was the father and which
the mother, the simpler and arguably better solution --which is
what has been implemented with this commit-- is to simply throw an
exception if either parent in any NuclearFamily has their gender
attribute set to missing.