Changed collapsedIndicatorSet from set to map
This commit represents the 2nd but still incomplete stage
in implementing a somewhat limited version of "collapsed"
functionality in Madeline 2.0.

In this commit, the collapsedIndicatorSet in PedigreeSet
has been changed from std::set to std::map<std::string,Individual *>
in order to keep track of and modify as needed the token individual
representing the collapsed group of individuals.

The Individual's collapsedCount is now incremented correctly but
gender and affection statuses within the group of collapsed are
not yet checked nor is there a display on the pedigree drawing of
the collapsed count.