Implement shading/coloring on Gender::MISSING
A collapsed group of affected individuals should be shaded just
the same as the uncollapsed based on affection status. However
a collapsed group may contain both males and females, necessitating
the use of the diamond symbol which is normally used for
Gender::MISSING individuals.

Prior to implementing the "collapsed" functionality, there was
no compelling need to shade Gender::MISSING individuals. That has
changed now. Therefore this commit represents the first step in
implementing shading/coloring on Gender::MISSING individuals who
are placeholders for a group of collapsed individuals.

(Partial implementaion, not done yet!)