Corrected y-positions when shading Gender::MISSING
Shading and positioning of "pie slice" segments to
show affection categories on icons was incorrect because
the diamond icon for Gender::MISSING is drawn slightly
higher so that all icon symbols —circle, square, and diamond—
will sit on a single baseline.

This has now been corrected so that the icon fill will also
be drawn based on the same slightly higher y coordinate center
for Gender::MISSING individuals.

Tested using both collapsed.data and ic_004 data.