Additions to support ART and unions w/o offspring
This commit adds two new options for unions without offspring.

In the current code, a heterosexual union without offspring is
indicated in the data via a surrogate "child" individual whose
IndividualID starts with a special character.

Previously, Madeline supported two options:

(1) An IndividualID starting with "^" indicated a union with no
    offspring, basically drawn like this:

    ▢─┬─○
      │
      ┴

(2) An IndividualID starting wtih "&" indicated a union with no
    offspring due to infertility, basically drawn like this:

    ▢─┬─○
      │
      ╧

This commit now adds the following two options:

(3) An IndividualID starting with "*" can be used to indicate
    a union with no vertical drop line at all, so basically
    drawn as follows:

    ▢───○

(4) Finally, an IndividualID starting with "%" can be used
    to indicate a union with no offspring but with the
    entwined rings symbol emphasizing that this is a union
    (and not, for example, some spurious drawing error):

    ▢─⚭─○

This final option, while currently only supporting heterosexual
unions, is envisaged to support homosexual unions in the future.