Sperm and ovum donor symbols now draw correctly
With this commit, Madeline now supports presentation of
sperm and ovum donors as parents of a child.

To use this feature, include a column called "Donor" in
the data file. Then, for any child who was born from a
sperm donor, mark an "S" or "sperm" in the "Donor" column
of the data table. For any child born from an ovum donor,
mark an "O" or "ovum" in the data column. For a child born
of both a sperm and ovum donor, mark "SO", "OS", "sperm_and_ovum"
or "ovum_and_sperm" in the data column.

BE SURE TO MARK THE "S", "O", OR "SO" ON THE CHILD, NOT ON
THE PARENTS! This is necessary, because those same parents may
also be the parents of natural or non-donor birth children.

When drawing the pedigree, Madeline will draw a sperm symbol
next to the father and an ovum symbol next to the mother in
place of the horizontal line connecting the parental pair, as
appropriate.

NOTA BENE: More work is still required to support the drawing
of ART pedigrees and gestational donors.