Fixing edge case bug when only one parent ID
This commit fixes an edge case bug where only one
parent ID is specified in the data file. In such cases,
Madeline creates a random ID for the missing parent.

If there are multiple children in a sibship, Madeline
makes the assumption that there was only one missing parent
— i.e. only one missing mother or one missing father — so that
all the sibs in the sibship will be full sibs and not half sibs.
This is, by the way, nothing more than an assumption. In order
to do this, once a virtual parent has been created for one of the
children, the program needs to be able to look up that parent as
the new virtual spouse of the a priori existing parent so that all
the other children will get that virtual parent.

There was an omission in Individual.cpp where the code always assumed
that a spouse would be found.