Complete rewrite of determineDescentTrees algor.
This commit adds a completely rewritten algorithm
for determining the so-called "original founders"
who are at the roots of all descent trees in Madeline
2.0.

("Original founders" are simply the parents of
children who are without grandparents).

The new algorithm is straightforward, much easier
to understand, and executes in a few loops without
resorting to recursion.

(In contrast the old algorithm resorted to recursion
but was buggy which resulted in sometimes completely
missing descent trees, leading to truncated drawings).

Now that certain previously undrawable pedigrees can
be drawn, at least a few additional bugs have been
revealed, leading to imperfect drawings. However, these
will be investigated in the coming days! :-)