Making Madeline smarter about missing individuals
"Identified missing parents" are parents who are identified
in the FATHER or MOTHER column but do not have an entry in
the data table.

In the past, this absence could lead to drawing errors or
crashes, especially if the second parent was missing (i.e.,
identified only by a missing value indicator, ".").

This commit insures that "identified missing parents" are
added and treated properly to avoid problems.