Allow binary bit patterns to specify quadrants
Passing the -q flag causes Madeline to shade
icon quadrants using data from a single column, the
first "Affected(_xxx)" column encountered.

Historically, data in this column needed
to have values between 0 to 15 inclusive.
Integer values in this range are then
interpreted according to their binary
bit patterns. For example, 9 in binary is
1001, so the first and fourth quadrants (counting
from the Northeast quadrant and moving counter-clockwise)
would be shaded black. A value of decimal ten, being
1010 in binary, would cause the 2nd (northwest) and 4th
(southeast) quadrants to be shaded.

This commit now allows the user to directly enter
bit patterns instead of trying to convert between
binary and decimal.  So if you want the first and
third quadrant shaded, just enter "0101" as your
value in the Affected_quadrant or similarly-named
column.  Note that to construct the pattern, you
still need to "start from the right", i.e. 0001
for the first quadrant, 0010 for the second quadrant,
0100 for the third, and 1000 for the fourth quadrant.