Revision of consang. & external conn. lines
This commit represents an in-progress revision of
how the consanguinous and external connector
lines are drawn.

This is being done in response to a bug report
and an accompanying test data set.

After the bug is completely fixed, the data set
will be added to the list of test data sets for
continuous integration testing.