Changes to drawSteppedConnector
This commit changes the new drawSteppedConnector
function to account for the fact that children are
always drawn with their mother in M2.

NOTA BENE: This is likely to be fixed again as some
test cases still don't pass while other test cases
are improved and some regressions have been eliminated.