Initial draft of new testing framework
This commit adds a new testing framework
that users and developers can run at any time
in their local (modern "unixy") environment.

After installing Madeline on your machine,
to run the tests just navigate to the
"testing_framework/results" directory and
run the "process.sh" script that is present
in that directory.

The process.sh script will extract all the
command lines documented for all of the test pedigrees
in the "index.html" file found in the main
testing_framework directory. After running Madeline,
the script will produce a simple PASS/DIFFER report
indicating which pedigrees are identical to those
in the expected set, and which differ.

For complete information, please read the
"testing_framework/index.html" web document.