1) Added symbol for mitochondrial donor & 2) test
(1) Mitochondrial Donor

This commit adds a symbol for a mitochondrial donor
even though Madeline is not yet capable of showing
so-called "3 parent" IVF/ART pedigrees. The addition
of this symbol, along with the existing "gestational
donor" symbol, are for future support of 3-parent
IVF/ART pedigrees.

(2) This commit also adds a "Donor" test pedigree
which shows the symbology that Madeline uses to show
donor situations.