Fixed edge case of incorrect placement of hz conn.
This commit fixes a rare edge case where the horizontal line
connecting a single offspring to his or her parents was misplaced.

It is often the case that a single offspring who is married is
drawn horizontally offset so that his or her vertical alignment is
directly below his or her leftmost parent.

In rare cases where said offspring actually participates in more
than one descent tree, it may happen that they have already been
drawn somewhere else on the pedigree, so we cannot use the
previously stored X & Y coordinates as a reference point.

If we instead use the startX and currentY within the current invocation
of the NuclearFamily::draw() command, then the connecting line will
always be drawn at the correct location.

In this commit, the test pedigree 0997_family.data has also been
modified to simply testing without changing anything else.