Fixed an omission in definition of "MDTJS" spouses
This commit fixes an omission in the definition of
"multiple descent tree joiner spouses" (MDTJS spouses).

The definition relies on detecting that a married pair have
no descent trees in common in the intersection set of their
descent tree vectors.

However the previous code failed to check whether an ordinary
founder spouse might not belong to any "other" descent tree and
therefore would have a descent tree vector of size zero, so the
intersection set would always yield zero, which is wrong.

Most multiple spouse pedigrees will now be drawn properly again
(as they had been prior to the addition of the MDTJS flag).

descent_tree_004 and descent_tree_005 will continue to be drawn
correctly with the corrected definition of MDTJS.

Only 0997_family —which in any case was problematic before too— will
continue to not be drawn correctly.