Missing parents now assigned "virtual_00n" IDs
With this commit, we no longer assign random ID labels
to missing parents. Instead, missing parents are now
given an ID with "virtual_" as the prefix and then simply
enumerated from 001 ... 999.

One reason for doing this was to be able to automate the
testing of the output against the expected output.

Secondly, it seems that IDs like "virtual_001", "virtual_002",
"virtual_003" are easier to read than the randomly-generated
identifiers used in the past.